Udpate: Udpate to make navbar re render when logged in

+++ src/components/navbar/navbar.tsx
@@ -1,8 +1,10 @@
 import "./navbar.css";
 
-import { Link } from "react-router-dom";
+import { Link, useLocation } from "react-router-dom";
 
 export default function Navbar() {
+  const location = useLocation();
+
   return (
     <nav className="FRE__Navbar">
       <div className="FRE__Navbar__Img">
@@ -11,41 +13,57 @@ export default function Navbar() {
         </Link>
       </div>
 
-      <ul>
-        <li>
-          <Link to="#">About FreeMe</Link>
-        </li>
-      </ul>
+      {(() => {
+        if (sessionStorage.getItem("uuid") === null) {
+          return (
+            <ul>
+              <li>
+                <Link to="#">About FreeMe</Link>
+              </li>
+            </ul>
+          );
+        }
+      })()}
 
       <div>
-        {sessionStorage.getItem("token") === null ? (
-          <>
-            <button>
-              <Link to="/login">
-                Login <i className="fa-solid fa-right-to-bracket"></i>
-              </Link>
-            </button>
-          </>
-        ) : (
-          <>
-            <button>
-              <div>
-                <Link to="/dashboard">
-                  Dashboard <i className="fa-solid fa-table-columns"></i>
+        {(() => {
+          if (
+            location.pathname === "/dashboard" ||
+            sessionStorage.getItem("uuid") !== null
+          ) {
+            return (
+              <>
+                <button>
+                  <Link to="/profile">Profile</Link>
+                </button>
+                <button>
+                  <div>
+                    <Link to="/dashboard">Dashboard</Link>
+                  </div>
+                </button>
+                <button
+                  onClick={() => {
+                    sessionStorage.clear();
+                  }}
+                >
+                  <Link to="/login">
+                    Logout <i className="fa-solid fa-right-from-bracket"></i>
+                  </Link>
+                </button>
+              </>
+            );
+          }
+
+          return (
+            <>
+              <button>
+                <Link to="/login">
+                  Login <i className="fa-solid fa-right-to-bracket"></i>
                 </Link>
-              </div>
-            </button>
-            <button
-              onClick={() => {
-                sessionStorage.clear();
-              }}
-            >
-              <Link to="/">
-                Logout <i className="fa-solid fa-right-from-bracket"></i>
-              </Link>
-            </button>
-          </>
-        )}
+              </button>
+            </>
+          );
+        })()}
       </div>
     </nav>
   );
